fix: add missing fields when writing to clickhouse

+++ pkg/data/clickhouse/evmrepo/transaction_row.go
@@ -34,6 +34,6 @@ type TransactionRow struct {
 	Input             string
 	Type              uint8
 	TransactionIndex  uint64
-	Success           uint8 // 1 for success, 0 for failure (from transaction receipt)
+	Success           uint64 // 1 for success, 0 for failure (from transaction receipt)
 	NumLogs           uint32
 }
